refactor(reporting): move reports to views/reporting/reports.py
Split saved-report CRUD (api_reports_list/get/create/update/delete),
sharing (api_reports_shares_get/set/delete, _is_report_owner,
_resolve_user, _shares_payload), the _preview_kind/_preview_summary
library-card helpers, and the share-target typeahead
(api_share_targets) out of nx_lib/views/reporting/__init__.py into
reports.py, part of beautify-phase-2a Task 3. __init__.py re-exports
_is_report_owner (the not-yet-moved schedule routes still call it),
_preview_kind, and _preview_summary, and delegates route registration
to reports.register_routes(app) -- which also registers
/api/reporting/share_targets, so that route now registers earlier
than before (no functional effect; Flask route registration order is
not path-sensitive here).

Retargeted the _preview_kind patch in
test_reports_list_skips_row_when_preview_kind_still_raises to
nx_lib.views.reporting.reports._preview_kind (api_reports_list calls
its own module's binding now). Verified with a break/revert probe.
